requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.16.0
requests==2.31.0

# AI/ML
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Email
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
emails==0.6.0

# Data Processing
pandas==2.1.4
numpy==1.24.4
python-dateutil==2.8.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Utilities
click==8.1.7
rich==13.7.0
schedule==1.2.1
jinja2==3.1.2
email-validator==2.1.0 